requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
Django==5.0.6
djangorestframework==3.15.1
flake8==7.0.0
importlib_metadata==7.1.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
Pygments==2.18.0
python-dotenv==1.0.1
sqlparse==0.5.0
tomli==2.0.1
zipp==3.19.0
